[android] consolidate app and service
