--- android/openthread_commissioner/app/.gitignore
@@ -1,1 +1,0 @@
-/build
