+++ android/openthread_commissioner/app/build.gradle
@@ -78,6 +78,11 @@ android {
     }
 }
 
+// This is for the local tests (src/test) to find the native libraries.
+tasks.withType(Test) {
+    systemProperty "java.library.path", "/usr/local/lib"
+}
+
 configurations.all {
     resolutionStrategy {
         force 'androidx.test:core:1.5.0'
@@ -85,12 +90,22 @@ configurations.all {
 }
 
 dependencies {
-    implementation project(':service')
+    implementation 'org.jetbrains:annotations:15.0'
+    def room_version = "2.2.5"
+    implementation "androidx.room:room-runtime:$room_version"
+    annotationProcessor "androidx.room:room-compiler:$room_version"
 
+    implementation fileTree(dir: "libs", include: ["*.jar"])
+
+    implementation 'com.google.guava:guava:31.1-jre'
     implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation "androidx.concurrent:concurrent-futures:1.1.0"
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.2'
+    implementation 'androidx.navigation:navigation-fragment:2.3.0'
+    implementation 'com.google.android.gms:play-services-vision:20.1.3+'
+    implementation 'com.google.android.gms:play-services-threadnetwork:16.0.0'
     implementation 'com.google.android.material:material:1.2.1'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.1'
-    testImplementation 'junit:junit:4.0'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    androidTestImplementation 'com.google.truth:truth:1.4.4'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
 }
