+++ android/openthread_commissioner/app/src/main/java/io/openthread/commissioner/app/CameraSourceView.java
@@ -26,7 +26,7 @@
  *    POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.openthread.commissioner.service;
+package io.openthread.commissioner.app;
 
 import android.Manifest;
 import android.content.Context;
