+++ android/openthread_commissioner/app/src/main/java/io/openthread/commissioner/app/NativeCommissionerLogger.java
@@ -26,26 +26,17 @@
  *    POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.openthread.commissioner.service;
+package io.openthread.commissioner.app;
 
-import android.app.Application;
-import android.content.Context;
+import android.util.Log;
+import io.openthread.commissioner.LogLevel;
+import io.openthread.commissioner.Logger;
 
-/** The Commissioner Service App for getting Application Context from anywhere. */
-public class CommissionerServiceApp extends Application {
-  private static CommissionerServiceApp instance;
-
-  public static CommissionerServiceApp getInstance() {
-    return instance;
-  }
-
-  public static Context getContext() {
-    return instance.getApplicationContext();
-  }
+public class NativeCommissionerLogger extends Logger {
+  private static final String TAG = "NativeCommissioner";
 
   @Override
-  public void onCreate() {
-    instance = this;
-    super.onCreate();
+  public void log(LogLevel level, String region, String msg) {
+    Log.d(TAG, String.format("[ %s ]: %s", region, msg));
   }
 }
