+++ android/openthread_commissioner/app/src/main/res/drawable/ic_done_icon.xml
