+++ android/openthread_commissioner/app/src/main/res/drawable/ic_error_icon.xml
