+++ android/openthread_commissioner/app/src/main/res/drawable/ic_network_selected_icon.xml
