+++ android/openthread_commissioner/app/src/main/res/drawable/ic_wireless_network_icon.xml
