+++ android/openthread_commissioner/app/src/main/res/layout/activity_commissioner.xml
