+++ android/openthread_commissioner/app/src/main/res/layout/fragment_fetch_credential_dialog.xml
