+++ android/openthread_commissioner/app/src/main/res/layout/fragment_meshcop.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.openthread.commissioner.service.CommissionerActivity">
+    tools:context=".MainActivity">
 
   <ProgressBar
       android:id="@+id/commissioning_progress"
