+++ android/openthread_commissioner/app/src/main/res/layout/fragment_network_password_dialog.xml
