+++ android/openthread_commissioner/app/src/main/res/layout/fragment_select_network.xml
