+++ android/openthread_commissioner/app/src/main/res/layout/network_list_item.xml
