+++ android/openthread_commissioner/app/src/main/res/values/styles.xml
