--- android/openthread_commissioner/service/.gitignore
@@ -1,3 +1,0 @@
-/build
-*.jar
-*.so
