--- android/openthread_commissioner/service/src/main/res/values/strings.xml
@@ -1,22 +1,0 @@
-<resources>
-    <string name="commissioning">Commissioning</string>
-    <string name="commissioner">Commissioner</string>
-    <string name="scan_device">Scan Device</string>
-    <string name="cancel">Cancel</string>
-    <string name="done">Done</string>
-    <string name="scan_instruction">Please scan the QR code on the device</string>
-    <string name="your_device">Your Device</string>
-    <string name="select_network">Select Network</string>
-    <string name="add_device">Add Device</string>
-    <string name="scan_qrcode_fragment_label">Scan QR Code</string>
-    <string name="select_network_fragment_label">Select Network</string>
-    <string name="password">password</string>
-    <string name="password_connect">connect</string>
-    <string name="password_cancel">cancel</string>
-    <string name="fetch_credential_done">done</string>
-    <string name="fetch_credential_cancel">cancel</string>
-    <string name="commissioning_fragment_label">Commissioning</string>
-    <string name="location_permission_missing_alert_title">Location permission missing</string>
-    <string name="location_permission_missing_alert_message">Location permission is required to scan BLE devices.</string>
-    <string name="location_permission_missing_alert_try_again">Try again</string>
-</resources>
